requirements.txt
absl-py==0.4.1
cycler==0.10.0
decorator==4.3.0
enum34==1.1.6
imageio==2.3.0
ipdb==0.10.3
ipython==4.0.0
ipython-genutils==0.2.0
kiwisolver==1.0.1
matplotlib==2.2.3
numpy==1.15.1
opencv-python==3.4.2.17
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
protobuf==3.6.1
ptyprocess==0.6.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
scipy==1.1.0
simplegeneric==0.8.1
six==1.11.0
tensorboardX==1.4
torchvision==0.2.1
traitlets==4.3.2
